% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covizerate.R
\name{covizerate}
\alias{covizerate}
\title{Create a covizerate visualization}
\usage{
covizerate(data, title = NULL)
}
\arguments{
\item{data}{a table of data}

\item{title}{chart title}
}
\description{
Create a covizerate visualization
}
