% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covizerate.R
\name{cvz_options}
\alias{cvz_options}
\title{Various options for the covizerate widget}
\usage{
cvz_options(covizerate, vline_stroke = "red", grad_fill = "#E8BF6A")
}
\arguments{
\item{covizerate}{An object of class 'covizerate'}

\item{vline_stroke}{Stroke color of the vertical line}

\item{grad_fill}{Fill color of the grad potion}
}
\value{

}
\description{
Various options for the covizerate widget
}
